#885 Implemented PersistenEntitySaver as the default implementation of contract IEntityActuator. Refactored CommonEntityDao to use PersistentEntitySaver.
